Bo2 focused update 1
Renamed all xnotifys
Added xmessage on dashboard launch

BO2
added auto shoot
added laser
added nigels xbo godmode fix
moved ipsoof to main
moved spoof rank to host only tab
tweaked serverinfo
shield check now works as intended
temporalily removed rebuilt scoreboard needs a lot of work
re-done esps

BO3
Improved aimbot

Minecraft
Added support for upcoming menu

Ghost
Added menu base

+++ KatBase/menu.cpp
@@ -213,7 +213,7 @@ void AddFloatMenuOptionToList(FloatMenuOption* option)
 	FloatMenuOption_List.push_back(option);
 }
 
-void AddFloat_List(int menuScroll)
+void AddFloat_List(float menuScroll)
 {
 	for each (FloatMenuOption * option in FloatMenuOption_List)
 	{
@@ -222,7 +222,7 @@ void AddFloat_List(int menuScroll)
 	}
 }
 
-void SubFloat_List(int menuScroll)
+void SubFloat_List(float menuScroll) 
 {
 	for each (FloatMenuOption * option in FloatMenuOption_List)
 	{
